requirements.txt
Django==4.1.1
djangorestframework==3.13.1
Markdown==3.4.1
django-filter==22.1
dj-rest-auth==2.2.5
